[Fix] check and load dependencies when needed
